% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_effect.R
\name{avg_effect}
\alias{avg_effect}
\title{Calculate average effects (the mean effect over the matrix support)}
\usage{
avg_effect(parm, matList, adj_positions, interaction_effects = list())
}
\arguments{
\item{parm}{the parameter}

\item{matList}{list of pairwise and spatial matrices}

\item{adj_positions}{positions within the adjacency matrix}

\item{interaction_effects}{list of interaction effects}
}
\value{
the average effects (the mean effect over the matrix support)
}
\description{
Calculate average effects (the mean effect over the matrix support)
}
\keyword{internal}
