% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_transform_param.R
\name{backward_transform_param}
\alias{backward_transform_param}
\title{Calculates the backward transformation of the parameter}
\usage{
backward_transform_param(param)
}
\value{
the parameter after applying the backward transformation
}
\description{
Calculates the backward transformation of the parameter
}
\keyword{internal}
