% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_transform_param_jacobian.R
\name{backward_transform_param_jacobian}
\alias{backward_transform_param_jacobian}
\title{Calculates the Jacobian of the backwards transformation}
\usage{
backward_transform_param_jacobian(param)
}
\value{
the Jacobian of the backwards transformation
}
\description{
Calculates the Jacobian of the backwards transformation
}
\keyword{internal}
