% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Dmat.R
\name{calc_Dmat}
\alias{calc_Dmat}
\title{Calculates the matrix D used in the quadratic minimization problem}
\usage{
calc_Dmat(matList_full)
}
\arguments{
\item{matList_full}{combined list of pairwise and spatial matrices}
}
\value{
the matrix D
}
\description{
Calculates the matrix D used in the quadratic minimization problem
}
\keyword{internal}
