% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dvec.R
\name{calc_dvec}
\alias{calc_dvec}
\title{Calculates the vector d used in the quadratic minimization problem}
\usage{
calc_dvec(S, matList_full)
}
\arguments{
\item{S}{the Pearson correlation matrix}

\item{matList_full}{combined list of pairwise and spatial matrices}
}
\value{
the vector d
}
\description{
Calculates the vector d used in the quadratic minimization problem
}
\keyword{internal}
