% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_neighbor_effect.R
\name{calc_mean_neighbor_effect}
\alias{calc_mean_neighbor_effect}
\title{Calculates the average correlation for the spatial effect}
\usage{
calc_mean_neighbor_effect(matList, beta, delta, adj_positions)
}
\arguments{
\item{matList}{list of pairwise and spatial matrices}

\item{beta}{autocorrelation parameter of the CAR model}

\item{delta}{linear effect of the CAR model correlation matrix}

\item{adj_positions}{positions within the adjacency matrix}
}
\value{
the average correlation for the spatial effect
}
\description{
Calculates the average correlation for the spatial effect
}
\keyword{internal}
