% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Sigma_opt_frob.R
\name{calc_Sigma_opt_frob}
\alias{calc_Sigma_opt_frob}
\title{Minimizes the Frobenius norm via quadratic optimization}
\usage{
calc_Sigma_opt_frob(matList_full, corY, edge_constraints = c())
}
\arguments{
\item{matList_full}{combined list of pairwise and spatial matrices}

\item{corY}{a correlation matrix estimate}

\item{edge_constraints}{a vector indicating which effect is constrained}
}
\value{
a list giving the optimal parameter and corresponding matrix
}
\description{
Minimizes the Frobenius norm via quadratic optimization
}
\keyword{internal}
