% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tilde_G_inv_partial_beta.R
\name{calc_tilde_G_inv_partial_beta}
\alias{calc_tilde_G_inv_partial_beta}
\title{Calculates the derivative of the inverse of G (the CAR model matrix)}
\usage{
calc_tilde_G_inv_partial_beta(M, A, beta)
}
\arguments{
\item{M}{matrix used in the definition of the CAR model}

\item{A}{matrix used in the definition of the CAR model}

\item{beta}{autocorrelation parameter of the CAR model}
}
\value{
the derivative of the inverse of G (the CAR model matrix)
}
\description{
Calculates the derivative of the inverse of G (the CAR model matrix)
}
\keyword{internal}
