% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_matList.R
\name{combined_matList}
\alias{combined_matList}
\title{Adds combined effects to the matList via the Hadamard product}
\usage{
combined_matList(matList, interaction_effects = NULL, check_redundancy = FALSE)
}
\arguments{
\item{matList}{a list of all matrices}

\item{interaction_effects}{list of vector pairs indicating interactions}
}
\value{
named list with all matList combinations and their positions
}
\description{
Adds combined effects to the matList via the Hadamard product
}
\keyword{internal}
