% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_matList_partial_der.R
\name{combined_matList_partial_der}
\alias{combined_matList_partial_der}
\title{Computes the derivative of the correlation matrix w.r.t. beta}
\usage{
combined_matList_partial_der(
  matList,
  link_matList,
  tilde_G_inv_partial_beta,
  interaction_effects = list()
)
}
\arguments{
\item{matList}{a list of all matrices}

\item{link_matList}{not in use (could be removed)}

\item{tilde_G_inv_partial_beta}{derivative of G w.r.t. beta}

\item{interaction_effects}{list of vector pairs indicating interactions}
}
\value{
the derivative of the correlation matrix w.r.t. beta
}
\description{
Computes the derivative of the correlation matrix w.r.t. beta
}
\keyword{internal}
