% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_marginal_cor.R
\name{compute_marginal_cor}
\alias{compute_marginal_cor}
\title{Computes non-positive-semidefinite approximation of correlation matrix}
\usage{
compute_marginal_cor(dataset)
}
\arguments{
\item{dataset}{n x d matrix (n = number of observations, d = dimension)}
}
\value{
non-positive-semidefinite approximation of correlation matrix
}
\description{
Computes non-positive-semidefinite approximation of correlation matrix
}
\keyword{internal}
