% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_from_standard_errors.R
\name{cor_from_standard_errors}
\alias{cor_from_standard_errors}
\title{Computes correlation matrix from a normalized dataset (=standard errors)}
\usage{
cor_from_standard_errors(varepsilon)
}
\arguments{
\item{varepsilon}{n x d matrix (n = number of observations, d = dimension)}
}
\value{
the correlation matrix
}
\description{
Computes correlation matrix from a normalized dataset (=standard errors)
}
\keyword{internal}
