% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Estimates the correlation matrix of the dataset}
\usage{
correlation_matrix(dataset, mean_estim, sd_estim)
}
\arguments{
\item{dataset}{n x d matrix (n = number of observations, d = dimension)}

\item{mean_estim}{estimate of the mean vector of the dataset}

\item{sd_estim}{estimate of the standard deviation vector of the dataset}
}
\value{
an estimate of the correlation matrix
}
\description{
Estimates the correlation matrix of the dataset
}
\keyword{internal}
