% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_D_der.R
\name{eta_D_der}
\alias{eta_D_der}
\title{Calculates the derivative of the spatial average effect}
\usage{
eta_D_der(
  parm,
  matList,
  adj_positions,
  interaction_effects = list(),
  index = 4
)
}
\arguments{
\item{parm}{the parameter}

\item{matList}{a list of the matrices (pairwise+spatial)}

\item{adj_positions}{the positions of the spatial effect (if embedded)}

\item{interaction_effects}{the interaction effect pairs (2-pair-name-vectors)}

\item{index}{position of the spatial effect within the list}
}
\value{
the derivative of the spatial average effect
}
\description{
Calculates the derivative of the spatial average effect
}
\keyword{internal}
