% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_information.R
\name{Fisher_information}
\alias{Fisher_information}
\title{Computes the Fisher information matrix}
\usage{
Fisher_information(adj_positions, parm, matList, interaction_effects = list())
}
\arguments{
\item{adj_positions}{positions of the spatial effect (if embedded)}

\item{parm}{the parameter}

\item{matList}{list of the matrices (pairwise+spatial)}

\item{interaction_effects}{list of pairwise effects (vectors of named pairs)}
}
\value{
the Fisher information matrix
}
\description{
Computes the Fisher information matrix
}
\keyword{internal}
