% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_transform_param.R
\name{forward_transform_param}
\alias{forward_transform_param}
\title{Transforms the parameter using a logit and inverse softmax}
\usage{
forward_transform_param(param)
}
\arguments{
\item{param}{the parameter}
}
\value{
the transformed parameter
}
\description{
Transforms the parameter using a logit and inverse softmax
}
\keyword{internal}
