% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frob_scalar_prod.R
\name{frob_scalar_prod}
\alias{frob_scalar_prod}
\title{Calculates the Frobenius inner product between to square matrices}
\usage{
frob_scalar_prod(A, B)
}
\arguments{
\item{A}{a square matrix}

\item{B}{a square matrix}
}
\value{
the Frobenius inner product
}
\description{
Calculates the Frobenius inner product between to square matrices
}
\keyword{internal}
