% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_M_A.R
\name{get_M_A}
\alias{get_M_A}
\title{Computes the matrices needed for the spatial effect}
\usage{
get_M_A(adj_matrix)
}
\arguments{
\item{adj_matrix}{the adjacency matrix of the spatial effect}
}
\value{
a list containing Ml and Al, which are used to compute the
         correlation matrix for the spatial effect
}
\description{
Computes the matrices needed for the spatial effect
}
\keyword{internal}
