% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradLogLikLogParm_02.R
\name{GradLogLikLogParm_02}
\alias{GradLogLikLogParm_02}
\title{Calculates the gradient of the function of the transformed parameter}
\usage{
GradLogLikLogParm_02(
  adj_positions,
  logParm,
  matList,
  dataset,
  interaction_effects = list()
)
}
\arguments{
\item{adj_positions}{positions of the spatial effect (if embedded)}

\item{logParm}{the transformed parameter}

\item{matList}{the list of matrices (pairwise + spatial)}

\item{dataset}{an n (observations) x d (dimensions) matrix}

\item{interaction_effects}{list of interaction effects (vectors of names)}
}
\value{
the gradient of the function of the transformed parameter
}
\description{
Calculates the gradient of the function of the transformed parameter
}
\keyword{internal}
