% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradLogLikParm_02.R
\name{GradLogLikParm_02}
\alias{GradLogLikParm_02}
\title{Calculates the gradient of the loglikelihood or the gradient of Sigma}
\usage{
GradLogLikParm_02(
  adj_positions,
  parm,
  matList,
  dataset,
  interaction_effects = list(),
  return_Sigma_der = FALSE
)
}
\arguments{
\item{adj_positions}{positions of the spatial effect (if embedded)}

\item{parm}{the parameter}

\item{matList}{the list of matrices (pairwise + spatial)}

\item{dataset}{an n (observations) x d (dimensions) matrix}

\item{interaction_effects}{list of interaction effects (vectors of names)}

\item{return_Sigma_der}{returns the gradient of Sigma if TRUE}
}
\value{
the gradient of the loglikelihood or the gradient of Sigma
}
\description{
Calculates the gradient of the loglikelihood or the gradient of Sigma
}
\keyword{internal}
