% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikLogParm_02.R
\name{LogLikLogParm_02}
\alias{LogLikLogParm_02}
\title{Computes (a translation of) the loglikelihood for the transformed parameter}
\usage{
LogLikLogParm_02(
  adj_positions,
  logParm,
  matList,
  dataset,
  interaction_effects = list()
)
}
\arguments{
\item{adj_positions}{positions of spatial effect (if embedded)}

\item{logParm}{the transformed parameter}

\item{matList}{the list of matrices (pairwise + spatial)}

\item{dataset}{an n (observations) x d (dimension) matrix}

\item{interaction_effects}{list of interaction effects (vectors of names)}
}
\value{
(a translation of) the loglikelihood
}
\description{
Computes (a translation of) the loglikelihood for the transformed parameter
}
\keyword{internal}
