% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikParm_02.R
\name{LogLikParm_02}
\alias{LogLikParm_02}
\title{Computes (a translation of) the loglikelihood}
\usage{
LogLikParm_02(
  adj_positions,
  parm,
  matList,
  dataset,
  interaction_effects = list(),
  mus = NULL,
  sigmas = NULL
)
}
\arguments{
\item{adj_positions}{positions of spatial effect (if embedded)}

\item{parm}{the parameter}

\item{matList}{the list of matrices (pairwise + spatial)}

\item{dataset}{an n (observations) x d (dimension) matrix}

\item{interaction_effects}{list of interaction effects (vectors of names)}

\item{mus}{the mean parameters}

\item{sigmas}{the standard deviation parameters}
}
\value{
(a translation of) the loglikelihood
}
\description{
Computes (a translation of) the loglikelihood
}
\keyword{internal}
