% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_support_distance.R
\name{mat_support_distance}
\alias{mat_support_distance}
\title{Computes a measure of distance between the support of two matrices}
\usage{
mat_support_distance(mat1, mat2)
}
\arguments{
\item{mat1}{a square matrix}

\item{mat2}{a square matrix}
}
\value{
a measure of distance between the support of two matrices
}
\description{
Computes a measure of distance between the support of two matrices
}
\keyword{internal}
