% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scov.R
\name{scov}
\alias{scov}
\title{Computes a structured estimator for covariance matrices}
\usage{
scov(
  pairwise_covariate_matrices,
  dataset,
  adj_matrix = NULL,
  mean_estim = NULL,
  sd_estim = NULL,
  grid_size = 100,
  parallelize = FALSE,
  ncores = 8,
  adj_positions = 1:nrow(adj_matrix),
  interaction_effects = list(),
  init = NULL,
  use_bootstrap = FALSE,
  num_bootstrap_iters = 100,
  semiparametric = FALSE,
  misspecification = FALSE,
  seed = 0,
  verbose = TRUE,
  joint_estimation = FALSE
)
}
\arguments{
\item{pairwise_covariate_matrices}{named list of square matrices}

\item{dataset}{the dataset given in matrix form}

\item{adj_matrix}{adjacency matrix of the spatial covariate}

\item{mean_estim}{mean vector estimate}

\item{sd_estim}{standard deviation vector estimate}

\item{grid_size}{grid-size for spatial effect}

\item{parallelize}{uses parallel-processing if TRUE}

\item{ncores}{number of cores for the parallelization}

\item{adj_positions}{positions within the adjacency matrix}

\item{interaction_effects}{list of interaction effects}

\item{init}{the initialization parameter vector}

\item{use_bootstrap}{uses bootstrapping if TRUE}

\item{num_bootstrap_iters}{number of bootstrap simulations}

\item{semiparametric}{computes the IVE if TRUE, the SCE else}

\item{misspecification}{computes the WSCE if TRUE, the WSCE else}

\item{seed}{a seed (can't be set to NULL)}

\item{verbose}{prints progress if TRUE}

\item{joint_estimation}{estimates everything jointly if TRUE,
uses a 2 step procedure if FALSE}
}
\value{
Returns a named list with the following elements:

         parm,             estimated parameters of pairwise, spatial effects
         average_effects,  average effects of the covariates
         corrmat_estim,    estimator of the correlation matrix
         covmat_estim,     estimator of the covariance matrix
         bic,              the Bayesian information criterion (BIC)
         lambda,           the asymptotically optimal weight of the WSCE
}
\description{
This function computes the WSCE, SCE or IVE estimator for large covariances
in the presence of pairwise and spatial covariates from
Metodiev et al. (2026).
}
\examples{

intercept = matrix(1,ncol=4,nrow=4)
X1 = rbind(c(1,1,1,0),c(1,1,1,0),c(1,1,1,0),c(0,0,0,1))
X2 = rbind(c(1,0,0,0),c(0,1,1,1),c(0,1,1,1),c(0,1,1,1))
covar_mats = list(intercept=intercept,X1=X1,X2=X2)
adj_matrix = rbind(c(0,1,0,0),c(1,0,0,0),c(0,0,0,1),c(0,0,1,0))
mean = rep(0,4)
sigma = 0.05*intercept+0.2*X1+0.2*X2+0.1*X2*X1+0.4*(diag(4) + adj_matrix)
diag(sigma) = 1
dataset = mvtnorm::rmvnorm(1000,mean=mean,sigma=sigma)
scov(covar_mats, dataset, adj_matrix,
interaction_effects=list(c("X1","X2")),
parallelize=FALSE,ncores=1)
}
\references{
Martin Metodiev. Marie Perrot-Dockès. Sarah Ouadah.
Bailey K. Fosdick. Stéphane Robin. Pierre Latouche. Adrian E. Raftery.
"A structured estimator for large covariance matrices in the presence of
pairwise and spatial covariates." Ann. Appl. Stat. 20 (2) 1736 - 1765,
June 2026.
}
