% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{test_func}
\alias{test_func}
\title{Bootstrap simulations for the WSCE}
\usage{
test_func(
  s,
  num_bootstrap_iters,
  seed,
  dimension,
  num_observations,
  pearson_mat,
  dataset,
  parm,
  pairwise_covariate_matrices,
  adj_matrix,
  grid_size,
  adj_positions,
  interaction_effects,
  verbose,
  joint_estimation
)
}
\arguments{
\item{s}{label of the sample}

\item{num_bootstrap_iters}{number of iterations}

\item{seed}{the seed}

\item{dimension}{the dimension of the data}

\item{num_observations}{size of the data}

\item{pearson_mat}{the pearson correlation matrix}

\item{dataset}{the dataset}

\item{parm}{the parameter estimated from the sce}

\item{pairwise_covariate_matrices}{the list of covariate matrices}

\item{adj_matrix}{adjacency matrix of the spatial effect}

\item{grid_size}{size of the grid for the spatial effect}

\item{adj_positions}{positions within the adjacency matrix}

\item{interaction_effects}{list of interaction effects}

\item{verbose}{prints progress if TRUE}

\item{joint_estimation}{estimates everything jointly if TRUE,
uses a 2 step procedure if FALSE}
}
\value{
Returns an array containing the simulated value of the SCE and
         the Pearson correlation matrix
}
\description{
This function is computed multiple times as a bootstrap sample
}
\keyword{internal}
