% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilde_G_inv.R
\name{tilde_G_inv}
\alias{tilde_G_inv}
\title{Computes the inverse of the correlation matrix of the CAR model}
\usage{
tilde_G_inv(
  M = NULL,
  A = NULL,
  beta,
  U_full = NULL,
  solve_U_full = NULL,
  solve_M_no_islands = NULL,
  eigen_real = NULL,
  return_U_D_M = FALSE
)
}
\arguments{
\item{M}{used to define the CAR correlation matrix}

\item{A}{used to define the CAR correlation matrix}

\item{beta}{autocorrelation parameter of the CAR model}

\item{U_full}{(optional) can be used to fasten computation}

\item{solve_U_full}{(optional) can be used to fasten computation}

\item{solve_M_no_islands}{(optional) can be used to fasten computation}

\item{eigen_real}{(optional) can be used to fasten computation}

\item{return_U_D_M}{(optoinal) can be used to fasten computation}
}
\value{
the inverse of the correlation matrix of the CAR model
}
\description{
Computes the inverse of the correlation matrix of the CAR model
}
\keyword{internal}
