% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_positive_definite.R
\name{to_positive_definite}
\alias{to_positive_definite}
\title{Maps the pairwise covariates to symmetric, positive definite matrices}
\usage{
to_positive_definite(pairwise_covariate_matrices, eig.tol = 1e-06)
}
\arguments{
\item{pairwise_covariate_matrices}{list of pairwise covariate matrices}

\item{eig.tol}{treat EV a as negative if (a / largest eigenvalue) <= eig.tol}
}
\value{
list of pairwise covariate matrices
}
\description{
Maps the pairwise covariates to symmetric, positive definite matrices
}
\keyword{internal}
