% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_LogLikParm_02.R
\name{true_LogLikParm_02}
\alias{true_LogLikParm_02}
\title{Computes the "true" (i.e., not translated) log-likelihood (needed for BIC)}
\usage{
true_LogLikParm_02(
  adj_positions,
  parm,
  matList,
  dataset,
  interaction_effects = list()
)
}
\arguments{
\item{adj_positions}{positions of spatial effect (if embedded)}

\item{parm}{the parameter}

\item{matList}{list of matrices (pairwise + spatial)}

\item{dataset}{an n (observations) x d (dimension) matrix}

\item{interaction_effects}{list of interaction effects (vector of names)}
}
\value{
the "true" (i.e., not translated) log-likelihood
}
\description{
Computes the "true" (i.e., not translated) log-likelihood (needed for BIC)
}
\keyword{internal}
